Changes to load correct metadata.

+++ src/main/java/ir/niopdc/policy/domain/policyversion/PolicyVersionKey.java
@@ -10,6 +10,6 @@ import lombok.ToString;
 @Setter
 @ToString
 public class PolicyVersionKey {
-    private Integer policyId;
+    private Byte policyId;
     private String version;
 }
